+++ NEWS.md
@@ -1,5 +1,8 @@
 # forcats 0.1.1.9000
 
+* `lvls_reorder()`, `fct_inorder()`, and `fct_infreq()` gain an `ordered`
+   argument, allowing you to override the existing "ordered" status (#54).
+
 * `fct_recode()` is now checks that each new value is of length 1 (#56).
 
 # forcats 0.1.1